back/requirements.txt
django
djangorestframework
markdown
django-filter
pygments
coreapi
django-cors-headers
psycopg2-binary
django-environ
uwsgi
celery
djangorestframework-camel-case
